+++ UI/ChallengeDetails/Tabs/Roster.lua
@@ -48,6 +48,7 @@ function MyDungeonsBook:RosterFrame_PartyMemberFrame_Create(rosterFrame, unitId,
 	self:RosterFrame_PartyMemberFrame_ClassIcon_Create(partyMemberFrame, unitId, challengeId);
 	self:RosterFrame_PartyMemberFrame_SpecIcon_Create(partyMemberFrame, unitId, challengeId);
 	self:RosterFrame_PartyMemberFrame_Race_Create(partyMemberFrame, unitId, challengeId);
+	self:RosterFrame_PartyMemberFrame_Covenant_Create(partyMemberFrame, unitId, challengeId);
 	self:RosterFrame_PartyMemberFrame_Equipment_Create(partyMemberFrame, unitId, challengeId);
 	return partyMemberFrame;
 end
@@ -113,7 +114,7 @@ Creates a frame with `unitId`'s race.
 @param[type=Frame] parentFrame
 @param[type=unitId] unitId
 @param[type=number] challengeId
-@return[type=Frame] specFrame
+@return[type=Frame] raceFrame
 ]]
 function MyDungeonsBook:RosterFrame_PartyMemberFrame_Race_Create(parentFrame, unitId, challengeId)
 	local challenge = self.db.char.challenges[challengeId];
@@ -124,6 +125,72 @@ function MyDungeonsBook:RosterFrame_PartyMemberFrame_Race_Create(parentFrame, un
 	end
 	raceFrame:SetWidth(150);
 	return raceFrame;
+end
+
+--[[--
+Creates a frame with `unitId`'s covenant.
+
+@param[type=Frame] parentFrame
+@param[type=unitId] unitId
+@param[type=number] challengeId
+]]
+function MyDungeonsBook:RosterFrame_PartyMemberFrame_Covenant_Create(parentFrame, unitId, challengeId)
+	local challenge = self.db.char.challenges[challengeId];
+	local covenantIconFrame = AceGUI:Create("InteractiveLabel");
+	parentFrame:AddChild(covenantIconFrame);
+	local unitCovenantInfo = challenge.players[unitId] and challenge.players[unitId].covenant;
+	if (not unitCovenantInfo.id) then
+		return;
+	end
+	local id = unitCovenantInfo.id;
+	local covenantMetaInfo = C_Covenants.GetCovenantData(unitCovenantInfo.id);
+	local suffix = self:GetIconTextureSuffix(30);
+	covenantIconFrame:SetText("|T" .. self:GetCovenantIconId(id) .. suffix .. "|t");
+	covenantIconFrame:SetCallback("OnEnter", function()
+		GameTooltip:SetOwner(covenantIconFrame.frame, "ANCHOR_NONE");
+		GameTooltip:SetPoint("BOTTOMLEFT", covenantIconFrame.frame, "BOTTOMRIGHT");
+		GameTooltip:AddLine(covenantMetaInfo.name, 1, 1, 1);
+		GameTooltip:Show();
+	end);
+	covenantIconFrame:SetCallback("OnLeave", function()
+		self:Table_Cell_MouseOut();
+	end);
+	covenantIconFrame:SetWidth(40);
+	local soulbindFrame = AceGUI:Create("Label");
+	soulbindFrame:SetText(unitCovenantInfo.soulbind.name);
+	soulbindFrame:SetWidth(60);
+	parentFrame:AddChild(soulbindFrame);
+	table.sort(unitCovenantInfo.soulbind.conduits, function (c1, c2)
+		return c1.row < c2.row;
+	end);
+	for _, conduit in pairs(unitCovenantInfo.soulbind.conduits) do
+		local conduitFrame = AceGUI:Create("InteractiveLabel");
+		local spellId = conduit.spellID;
+		local icon = conduit.icon;
+		if (spellId == 0 and conduit.conduitID ~= 0) then
+			spellId = C_Soulbinds.GetConduitSpellID(conduit.conduitID, conduit.conduitRank);
+		end
+		if (conduit.conduitID ~= 0) then
+			icon = select(3, GetSpellInfo(spellId));
+		end
+		conduitFrame:SetText("|T".. icon .. suffix .."|t");
+		conduitFrame:SetWidth(40);
+		conduitFrame:SetCallback("OnEnter", function()
+			GameTooltip:SetOwner(conduitFrame.frame, "ANCHOR_NONE");
+			GameTooltip:SetPoint("BOTTOMLEFT", conduitFrame.frame, "BOTTOMRIGHT");
+			if (conduit.conduitID == 0) then
+				GameTooltip:SetSpellByID(spellId);
+			else
+				GameTooltip:SetHyperlink(C_Soulbinds.GetConduitHyperlink(conduit.conduitID, conduit.conduitRank));
+			end
+			GameTooltip:Show();
+		end);
+		conduitFrame:SetCallback("OnLeave", function()
+			self:Table_Cell_MouseOut();
+		end);
+		parentFrame:AddChild(conduitFrame);
+	end
+	return covenantIconFrame;
 end
 
 --[[--